+++ stac_fastapi/core/stac_fastapi/core/database_logic.py
@@ -96,7 +96,13 @@ ES_MAPPINGS_DYNAMIC_TEMPLATES = [
     },
     # Default all other strings not otherwise specified to keyword
     {"strings": {"match_mapping_type": "string", "mapping": {"type": "keyword"}}},
-    {"numerics": {"match_mapping_type": "long", "mapping": {"type": "float"}}},
+    {"long_to_double": {"match_mapping_type": "long", "mapping": {"type": "double"}}},
+    {
+        "double_to_double": {
+            "match_mapping_type": "double",
+            "mapping": {"type": "double"},
+        }
+    },
 ]
 
 ES_ITEMS_MAPPINGS = {